refactor(ui): relocate shared composable and util out of FormLayout
Move useDoctypeMeta to src/composables and scopedRegistry to src/utils,
surfacing both from the root barrel; add a ./utils subpath export.
