+++ ui/src/index.ts
@@ -1,5 +1,5 @@
-export * from './components/Link'
-export * from './components/Phone'
+export * from "./components/Link";
+export * from "./components/Phone";
 
 // FormLayout: only the headline surface is re-exported from the root barrel.
 // The full API (formatting utils, field-type registry, extra composables and
@@ -14,7 +14,9 @@ export type {
   FieldComponentProps,
   FieldComponentEmits,
 } from "./components/FormLayout";
+export { useDoctypeMeta } from "./composables/useDoctypeMeta";
+export type { UseDoctypeMeta, DoctypeMeta } from "./composables/useDoctypeMeta";
+export * from "./utils";
 export * from "./components/Grid";
 export * from "./components/Link";
 export * from "./components/TableMultiSelect";
-
