+++ ui/src/utils/scopedRegistry.ts
