+++ ui/src/utils/tests/scopedRegistry.test.ts
